refactor: data loader에서 train과 inference 패키지 분리
